Add Authorization

+++ backend/Routes/postsRoutes.js
@@ -37,6 +37,7 @@ router.post("",authCheck ,multer({storage: storage}).single('image'),(req, res, 
     title: req.body.title,
     content: req.body.content,
     imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   console.log(post);
   post.save().then(resultPost => {
@@ -61,11 +62,16 @@ router.put("/:id",authCheck,multer({storage: storage}).single('image'),(req,res,
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    console.log(result);
-    res.status(200).json({message: "Post Update Sucsseful"});
+  Post.updateOne({_id: req.params.id , creator: req.userData.userId}, post).then(result => {
+    if(result.n > 0) {
+      res.status(200).json({message: "Post Update Sucsseful"});
+    }else {
+      res.status(401).json({message: "User Not Auth"})
+    }
+
   })
 })
 
@@ -104,9 +110,13 @@ router.get("", (req, res, next) => {
 });
 
 router.delete("/:id",authCheck,(req,res,next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({message: "Post Deleted"});
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if(result.n) {
+
+      res.status(200).json({message: "Post Deleted"});
+    }else {
+      res.status(401).json({message: "User Not Auth"})
+    }
   });
 });
